Move initiate billing batch service
The initiate service deals with creating the charging module bill run, then merging its result with the WRLS billing batch record we are trying to create. It also creates a new billing batch event (though goodness knows why!)

Again, these are all things we need to do irrespective of the billing batch type so we make it generic and move it out of `supplementary/`.
